+++ hTech/BoundingPolygon.h
@@ -15,7 +15,7 @@ public:
 
 	BoundingPolygon(Transform& transform, float& rotation, const std::vector<Vector2> points, const int& pointCount);
 
-	virtual void Update(double DeltaTime);
+	virtual void Update(float DeltaTime);
 	virtual void Render(SDL_Renderer& renderer);
 
 	virtual Vector2 FindFurthestPoint(Vector2 direction) const;
